Refactor displayThread and displayForum methods to remove params and allow nulls
Removed forum parameters from #displayThread (don't need them) and added a
postJump parameter.
Made every parameter optional except IDs, so calls aren't explicitly requesting
page 1

+++ Awful.apk/src/main/java/com/ferg/awfulapp/ThreadDisplayFragment.java
@@ -374,7 +374,7 @@ public class ThreadDisplayFragment extends AwfulFragment implements SwipyRefresh
 			AwfulURL aLink = AwfulURL.parse(aUrl);
 			switch (aLink.getType()) {
 				case FORUM:
-					displayForum(aLink.getId(), aLink.getPage());
+					displayForum((int) aLink.getId(), (int) aLink.getPage());
 					break;
 				case THREAD:
 					if (aLink.isRedirect()) {
